+++ packages/interweave-emoji-picker/CHANGELOG.md
@@ -3,6 +3,8 @@
 * Added a new `disableGroups` prop to `Picker`, which will hide the tabs and disable group headers.
   * Emojis will now be grouped into a "none" group.
   * Commonly used emojis and searching are still available in this mode.
+* Added new `blacklist` and `whitelist` props to `Picker`, both of which accept an
+  array of hexcodes.
 
 #### 🛠 Internal
 * Commonly used emojis now store the `hexcode` instead of the `unicode` character.